Fixed UI UX

+++ client/src/page/center/homebase/TableData.jsx
@@ -16,16 +16,21 @@ const TableData = ({ setDetail }) => {
     useDeleteHomebaseMutation();
 
   const deleteHandler = (id) => {
-    toast.promise(
-      deleteHomebase(id)
-        .unwrap()
-        .then((res) => res.message),
-      {
-        loading: "Memproses...",
-        success: (message) => message,
-        error: (err) => err.data.message,
-      }
+    const confirm = window.confirm(
+      "Apakah anda yakin akan menghapus seluruh data satuan ini?"
     );
+    if (confirm) {
+      toast.promise(
+        deleteHomebase(id)
+          .unwrap()
+          .then((res) => res.message),
+        {
+          loading: "Memproses...",
+          success: (message) => message,
+          error: (err) => err.data.message,
+        }
+      );
+    }
   };
 
   useEffect(() => {
@@ -45,39 +50,36 @@ const TableData = ({ setDetail }) => {
       page={page}
       setPage={setPage}
       setLimit={setLimit}
-      setSearch={setSearch}
-    >
-      <table className="table table-bordered table-striped table-hover">
+      setSearch={setSearch}>
+      <table className='table table-bordered table-striped table-hover'>
         <thead>
           <tr>
-            <th className="text-center">No</th>
-            <th className="text-center">_id</th>
-            <th className="text-center">Satuan Pendidikan</th>
-            <th className="text-center">Aksi</th>
+            <th className='text-center'>No</th>
+            <th className='text-center'>_id</th>
+            <th className='text-center'>Satuan Pendidikan</th>
+            <th className='text-center'>Aksi</th>
           </tr>
         </thead>
         <tbody>
           {homebase?.map((item, i) => (
             <tr key={item.id}>
-              <td className="text-center align-middle">
+              <td className='text-center align-middle'>
                 {(page - 1) * limit + i + 1}
               </td>
-              <td className="text-center align-middle">{item.id}</td>
-              <td className="text-center align-middle">{item.name}</td>
+              <td className='text-center align-middle'>{item.id}</td>
+              <td className='text-center align-middle'>{item.name}</td>
               <td>
-                <div className="d-flex gap-2 justify-content-center">
+                <div className='d-flex gap-2 justify-content-center'>
                   <button
-                    className="btn btn-warning"
-                    onClick={() => setDetail(item)}
-                  >
-                    <i className="bi bi-pencil-square"></i>
+                    className='btn btn-sm btn-warning'
+                    onClick={() => setDetail(item)}>
+                    <i className='bi bi-pencil-square'></i>
                   </button>
                   <button
-                    className="btn btn-danger"
+                    className='btn btn-sm btn-danger'
                     disabled={isLoading}
-                    onClick={() => deleteHandler(item.id)}
-                  >
-                    <i className="bi bi-folder-minus"></i>
+                    onClick={() => deleteHandler(item.id)}>
+                    <i className='bi bi-folder-minus'></i>
                   </button>
                 </div>
               </td>
